CheckLogin Bit
Login is much more efficient and does a thorough check of the user type
to make sure the right users are allowed access to the correct profiles.

The login bit will be checked to '1' if the user is logged in. I haven't
made it return to 0 on logout yet so you will have to do that manually
for now.